Standardise on "MCP2221" rather than "MCP2221A"
I'd used both terms, so this is just for consistency.

The MCP2221A specifically is the more commonly available chip but they
work in the same way, other than the A allowing for higher UART speeds
and fixing the MCP2221 errata about internal UART register
(non-)initialisation causing the USB-UART bridge not to work.

I'll need to make sure to at least mention the MCP2221A in the readme so
people can find it while searching(!)