Fix description of clock output
The datasheet description of the clock output divider in the USB HID
commands section did not align with the description in register 1-2.

It also describes the chip as having a 48 MHz clock, but the MCP2221 has
a 12 MHz clock (as noted in several places in the datasheet).

I've edited the chip settings field description to align with that of
the register (2 bit duty cycle setting, 3 bit divider), but it's still
just a u8. Probably it would be beneficial to split this up into two
fields, each with its own type, as even the divider value is not
strictly a divider (0b010 -> 12MHz / 1 ; 0b001 -> 12 MHz * 2).