Enable DAC output via SRAM settings
This also entailed moving several types out of the flash_data module
because they are reused for reading and writing the SRAM settings.

Following teasting with the device, I've also changed the VRM/VDD
handling, by using 1 = VRM in all cases. This matches the register
description in the datasheet, but disagrees with two command listings
where it appears Microchip have listed 0 = VRM by mistake.

This also allowed me to unify the DAC & ADC voltage reference sources
into a single type. This should be further consolidated so that a VRM
reference level is only needed when VRM is in use.

There are some oddities with the DAC output that I should document for
users. With a supply voltage of 3.3V, the output is not linear, and in
none of the voltage modes does the highest output value (31) produce a
voltage higher than 70% of nomimal.

I've also added support for resetting the device, which drops the
driver.

And I've fixed a bug where the VID was overwritten with the PID in
flash. At some point both got set to zero, and I've yet to diagnose
whether that is a problem in my code, or if it resulted from the use of
another MCP2221 driver to check something (EasyMCP2221).

For now I've disabled setting the VID or PID, and just write the
default values.