Remove flawed WriteCommandData trait
Really, it was only used to avoid direct access to the write buffer in
the UsbReport struct. The problem, now I'm working on the SRAM settings,
is that some structs useful for both flash and SRAM settings need to be
written to the command buffers differently depending on which settings
are being altered (flash or SRAM).