hal: i2c: implement i2c write
Checked up to 256-byte writes, so a lot fewer than 65,535 (I'll get to
it!) but shows the chunking is working properly.

Also checking to see if a device is there (if it ACKs its address),
which is just a 0-length write followed by checking the address ACK
in the status output.

I specifically prevent 0-length "normal" writes because for the MCP2221
if the device isn't there you have to clean up by cancelling the
transfer, otherwise the next time you try to do an I2C transfer you'll
just get engine errors and it's not necessarily obvious why.

This way, people who want to do 0-length writes to find a device can
just the specific function and not have to worry about clean-up. I
should recreate the scan_bus example from mcp2221-rs to show this
usage.