hal: i2c: handle missing targets in read & write
With the previous commit, it was possible to call the i2c_read and
i2c_write functions and leave the MCP2221 I2C engine in an error state,
which could be resolved by separately cancelling the I2C transfer. Now
we check we received an address ACK after the first write or after the
read submit. If there was no ACK, we cancel the transfer and return an
error to the caller.